+++ app/components/primer/heading_component.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Primer
-  class HeadingComponent < Primer::Beta::Heading
-    status :deprecated
-  end
-end