+++ lib/primer/deprecations.rb
@@ -9,14 +9,8 @@ module Primer
       "Primer::Alpha::AutoComplete::Item" => "Primer::Beta::AutoComplete::Item",
       "Primer::BlankslateComponent" => "Primer::Beta::Blankslate",
       "Primer::BoxComponent" => "Primer::Box",
-      "Primer::CloseButton" => "Primer::Beta::CloseButton",
       "Primer::CounterComponent" => "Primer::Beta::Counter",
-      "Primer::DetailsComponent" => "Primer::Beta::Details",
       "Primer::DropdownMenuComponent" => nil,
-      "Primer::FlexComponent" => nil,
-      "Primer::FlexItemComponent" => nil,
-      "Primer::HeadingComponent" => "Primer::Beta::Heading",
-      "Primer::HiddenTextExpander" => "Primer::Alpha::HiddenTextExpander",
       "Primer::IconButton" => "Primer::Beta::IconButton",
       "Primer::Tooltip" => "Primer::Alpha::Tooltip"
     }.freeze