+++ lib/primer/forms/dsl/check_box_input.rb
@@ -8,7 +8,7 @@ module Primer
         DEFAULT_SCHEME = :boolean
         SCHEMES = [DEFAULT_SCHEME, :array].freeze
 
-        attr_reader :name, :label, :value, :unchecked_value, :scheme
+        attr_reader :name, :label, :value, :unchecked_value, :scheme, :nested_form_block, :nested_form_arguments
 
         def initialize(name:, label:, value: nil, unchecked_value: nil, scheme: DEFAULT_SCHEME, **system_arguments)
           raise ArgumentError, "Check box scheme must be one of #{SCHEMES.join(', ')}" unless SCHEMES.include?(scheme)
@@ -22,10 +22,17 @@ module Primer
           @scheme = scheme
 
           super(**system_arguments)
+
+          yield(self) if block_given?
         end
 
         def to_component
           CheckBox.new(input: self)
+        end
+
+        def nested_form(**system_arguments, &block)
+          @nested_form_arguments = system_arguments
+          @nested_form_block = block
         end
 
         def type