+++ test/components/alpha/nav_list_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "test_helper"
+require "components/test_helper"
 
 module Primer
   module Alpha
@@ -64,6 +64,34 @@ module Primer
         refute_selector ".ActionListItem--navActive", text: "Item 2"
       end
 
+      class CurrentPageItem < Primer::Alpha::NavList::Item
+        def initialize(current_page_href:, **system_arguments)
+          @current_page_href = current_page_href
+          super(**system_arguments)
+        end
+
+        private
+
+        def current_page?(url)
+          url == @current_page_href
+        end
+      end
+
+      def test_item_can_be_selected_by_current_page
+        current_page_href = "/item2"
+
+        render_inline(Primer::Alpha::NavList.new) do |c|
+          c.with_section(aria: { label: "Nav list" }) do |section|
+            # use CurrentPageItem instead of a mock since the API supports it
+            section.with_item(component_klass: CurrentPageItem, label: "Item 1", href: "/item1", current_page_href: current_page_href)
+            section.with_item(component_klass: CurrentPageItem, label: "Item 2", href: "/item2", current_page_href: current_page_href)
+          end
+        end
+
+        refute_selector ".ActionListItem--navActive", text: "Item 1"
+        assert_selector ".ActionListItem--navActive", text: "Item 2"
+      end
+
       def test_max_nesting_depth
         error = assert_raises(RuntimeError) do
           render_inline(Primer::Alpha::NavList.new) do |c|
@@ -78,6 +106,27 @@ module Primer
         end
 
         assert_equal "Items can only be nested 2 levels deep", error.message
+      end
+
+      def test_show_more_item
+        render_preview(:show_more_item)
+
+        assert_selector("#ActionList--showMoreItem", visible: false, text: "Show more")
+      end
+
+      def test_disallow_subitems_and_trailing_action
+        error = assert_raises(RuntimeError) do
+          render_inline(Primer::Alpha::NavList.new) do |c|
+            c.with_section(aria: { label: "List" }) do |section|
+              section.with_item(label: "Level 1", href: "/level1") do |item|
+                item.with_item(label: "Level 2", href: "/level2")
+                item.with_trailing_action(icon: :megaphone, aria: { label: "Action" })
+              end
+            end
+          end
+        end
+
+        assert_equal "Cannot render a trailing action for an item with subitems", error.message
       end
     end
   end