+++ test/components/component_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "test_helper"
+require "components/test_helper"
 require_relative "../../lib/primer/view_components/statuses"
 
 class PrimerComponentTest < Minitest::Test
@@ -11,6 +11,13 @@ class PrimerComponentTest < Minitest::Test
     [Primer::Alpha::SplitButton, {}],
     [Primer::Beta::IconButton, { icon: :star, "aria-label": "Star" }],
     [Primer::Beta::Button, {}],
+    [Primer::Alpha::SegmentedControl, {
+      full_width: false
+    }, proc { |component|
+      component.with_item(label: "Button", selected: true)
+      component.with_item(label: "Button")
+    }],
+    [Primer::Alpha::SegmentedControl::Item, { label: "Button" }],
     [Primer::Alpha::Layout, {}, proc { |component|
       component.main(tag: :div) { "Foo" }
       component.sidebar(tag: :div) { "Bar" }
@@ -71,9 +78,7 @@ class PrimerComponentTest < Minitest::Test
     end],
     [Primer::Dropdown::Menu, {}],
     [Primer::DropdownMenuComponent, {}],
-    [Primer::FlexComponent, {}],
     [Primer::Beta::Flash, {}],
-    [Primer::FlexItemComponent, { flex_auto: true }],
     [Primer::Beta::Heading, { tag: :h1 }],
     [Primer::Alpha::HiddenTextExpander, { "aria-label": "No action" }],
     [Primer::LabelComponent, {}],
@@ -109,11 +114,7 @@ class PrimerComponentTest < Minitest::Test
       "Primer::Alpha::ActionList::Item",
       "Primer::Alpha::ActionList::Separator",
       "Primer::Alpha::NavList::Section",
-      "Primer::HiddenTextExpander",
-      "Primer::HeadingComponent",
-      "Primer::CloseButton",
       "Primer::CounterComponent",
-      "Primer::DetailsComponent",
       "Primer::Component",
       "Primer::OcticonsSymbolComponent",
       "Primer::Content",